updating columns, changed to bigint

+++ src/main/resources/db/migrations/changes/18112019_create_table_users.sql
@@ -1,6 +1,6 @@
 create table users
 (
-    id              serial primary key not null,
+    id              bigint primary key not null,
     activation_code varchar(255),
     active          boolean,
     email           varchar(100),
